+++ slack_bot/slash_commands/__init__.py
@@ -1,0 +1,1 @@
+from api import router as slash_commands_router
